Refactor nomad.go
into multiple files to simplify the bloated file by introducing multiple sub-functionality handlers.
